requirements.txt
tensorflow-cpu
fastapi
opencv-python-headless
python-multipart
protobuf
uvicorn
python-dotenv==1.0.0
numpy
pymongo==4.5.0
uvicorn
